Compiling Lab 5 code without any other additions

+++ Lab5/ADC.h
@@ -2,4 +2,4 @@
  
 	void ADC_Init(unsigned char channelNum); 
 	unsigned long ADC_In(void); //Using Busy wait
-	void ADC_Collect(unsigned int channelNum, unsigned int fs, void (*ADCtask) (unsigned long)); //using interrupts
+	void ADC_Collect(unsigned int channelNum, unsigned int fs, void (*ADCtask) (unsigned short)); //using interrupts
